LayoutTests/http/tests/resourceLoadStatistics/website-data-removal-for-site-navigated-to-with-link-decoration-FilterLinkDecorationOnNavigationEnabled-expected.txt

Check that script-writeable website data gets removed after a navigation with link decoration from a prevalent resource.

Before deletion: Client-side cookie exists.
Before deletion: Client-side persistent cookie exists with 24-hour capped max-age: true
Before deletion: HttpOnly cookie exists.
Before deletion: Regular server-side cookie exists.
Before deletion: LocalStorage entry does exist.
Before deletion: IDB entry does exist.

Before deletion: Client-side cookie exists.
Before deletion: Client-side persistent cookie exists with 24-hour capped max-age: true
Before deletion: HttpOnly cookie exists.
Before deletion: Regular server-side cookie exists.
Before deletion: LocalStorage entry does exist.
Before deletion: IDB entry does exist.
After deletion: Client-side cookie does not exist.
After deletion: Client-side persistent cookie does not exist.
After deletion: HttpOnly cookie exists.
After deletion: Regular server-side cookie exists.
After deletion: LocalStorage entry does not exist.
After deletion: IDB entry does not exist.


Resource load statistics:

Registrable domain: 127.0.0.1
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    TopFrameUniqueRedirectsFrom:
        localhost
    TopFrameLinkDecorationsFrom:
        localhost
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 1
Registrable domain: 3rdpartytestwebkit.org
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: Yes
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: localhost
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    TopFrameUniqueRedirectsTo:
        127.0.0.1
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: Yes
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
